Fix NULL ptr dereference with bad type
Noted by Alexey Dobriyan, who also sent in a small test-case.

When evaluating to a bad type, just use "&bad_ctype" rather than
returning NULL.  That will also mean that we won't be re-evaluating that
expression over and over again - we'll just mark the expr->ctype to be
bad, once and for all.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>