[PATCH] Parse and track multiple contexts by expression
sparse currently only tracks one global context for __context__ and
__attribute__((context)).

This adds support for parsing an additional argument to each of these
which gives a context expression.  For __attribute__((context)), store
each context attribute as a separate context structure containing the
expression, the entry context, and the exit context, and keep a list of
these structures in the ctype.  For __context__, store the context
expression in the context instruction.  Modify the various frontends to
adapt to this change, without changing functionality.

This change should not affect parsing of programs which worked with
previous versions of sparse, unless those programs use comma expressions
as arguments to __context__ or __attribute__((context)), which seems
highly dubious and unlikely.  sparse with -Wcontext generates identical
output with or without this change on Linux 2.6.18-rc4.

Signed-off-by: Josh Triplett <josh@freedesktop.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>