Generate and install a pkg-config file.  Add DESTDIR support to Makefile.
The Makefile now generates and installs a pkg-config file, sparse.pc, to
provide information about the installed Sparse library and header files.  Use
`pkg-config --cflags sparse` to get the include path, and `pkg-config --libs
sparse` to get the library path and library.

To help packagers of Sparse, also add support for the DESTDIR variable to the
Makefile.  DESTDIR allows you to target a particular prefix (such as /usr),
but install to another directory (such as debian/tmp), usually for the
purposes of subsequently constructing a package from that directory.
Previously, just setting PREFIX would work for that, but with the new
pkg-config file, the Makefile needs to know the real prefix.

Signed-off-by: Josh Triplett <josh@freedesktop.org>