Sparse-based Ctags implementation
Here is my attempt to write a Ctags base on sparse.
It depends on the result from actually preprocessing the C source file.
It can handle the tags from macro expansion.

It works well enough for me to locate "int32_t" and "clear_token_alloc".

This is the updated ctags that actually generates output to a tags file.

Signed-Off-By: Christopher Li <sparse@chrisli.org>
Signed-Off-By: Josh Triplett <josh@freedesktop.org>