Use GCC format and sentinel attributes on appropriate functions
Expose the FORMAT_ATTR portability macro in lib.h, and use it on the various
printf-like functions in sparse.

Add a new SENTINEL_ATTR portability macro for the GCC sentinel attribute, and
use it on match_idents in parse.c.

match_oplist in expression.c should use SENTINEL_ATTR, but GCC does not accept
an integer 0 as a sentinel, only a pointer 0 like NULL.

Signed-off-by: Josh Triplett <josh@freedesktop.org>