Parse asm after a label as a statement, not an attribute
Commit aec53c938c34c47cdbdd6824552e0f2a5104b1cb, "handle label attributes",
caused sparse to parse asm after a label as an attribute, not a statement.
Fix this by adding a new allow_asm flag to handle_attributes, and passing 0
when parsing label attributes.  Expand validation/asm-volatile.c to include
the test case that demonstrated this bug.

Thanks to Randy Dunlap for reporting the problem.

Signed-off-by: Josh Triplett <josh@freedesktop.org>