Improve error message if using a member of an incomplete struct or union
sparse should distinguish two cases.  One is when the structure or union
is declared and the member is not found.  Another is when the structure
or union is not yet declared.  In the later case, the message should
indicate that the type is incomplete, rather than complain that the
particular member is not found.

Mention the incomplete type like gcc does, but provide the member name
too, it may be useful.

Signed-off-by: Pavel Roskin <proski@gnu.org>