[PATCH] fix the sanity check in evaluate_ptr_sub()
We are checking the wrong thing - the arguments have already
degenerated, so we won't see SYM_FN there (or SYM_ARRAY, for
that matter).  That check should be done to the target of
pointer, not to the pointer itself and it should simply check
for SYM_FN - that's what we'll get there both from decayed
functions _and_ pointers to functions.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>