Fix test-suite to handle stdout and stderr separately, and fix up tests
test-suite merged stdout and stderr, which relied on the ordering of data from
the two streams in the merged stream.  This made test-suite frequently fail on
tests with both output and errors, when the ordering didn't happen to match
what the test assumed.  Handle each of the streams separately, and update the
tests accordingly.

Also clean up the output of "test-suite format" to avoid outputting values
equal to the defaults.

Signed-off-by: Josh Triplett <josh@freedesktop.org>