Fix type mismatches with incomplete types
If I put the following in a public header:

   struct foo;
   typedef struct foo *Foo;

   void func (Foo f);

and the definition of struct foo in a private header:

   struct foo { int bar; }

then I get a sparse warning (different base type for argument 1) when
I compile the implementation file:

   #include "public.h"
   #include "private.h"

   void func (Foo f) { ... }

i.e. sparse doesn't realise that the incomplete structure definition
in the function  prototype refers to the same type as the complete
structure definition in the function definition.

*I think* that the patch fixes this - it silences the error - but I
don't know enough about sparse to know whether it's correct (or
whether it silences other legitimate errors, for example).

Signed-off-by: Geoff Johnstone <geoff.johnstone@googlemail.com>