Fix use of invalid file descriptor
tokenize_buffer() calls setup_stream() with a file descriptor argument of -1.
This number makes it all the way into nextchar_slow(), where -1 is used as an
actual file descriptor when calling read().

Check for the -1 before calling read() if the buffer is empty.

(By the way, that read() there is probably missing some additional error
 handling. An EINTR at the right moment...)

Signed-off-by: Vegard Nossum <vegardno@ifi.uio.no>