This is the start of the smatch_extra stuff.
The eventual goal is to try track the possible values of all the variables.

Currently the only use is to tell when we know a pre-loop condition is
true the first time.

Sometimes it is easy.
for (i = 0 ; i < 100 ; i++) { ...

We know that i < 100 the first time through the loop.


Signed-off-by: Dan Carpenter <dcarpenter@usbsys.site>