Try again removing states from pools.
There was a patch that just deleted a state from the pool but I had to
revert it because on some paths a state is part of the pool and some it's
not.  The new code looks in cur_slist to see what pools a state is in using
is_currently_in_pool().

That was confusing as pants...  :/

Signed-off-by: Dan Carpenter <error27@gmail.com>