if (a || b) { a->foo; should print an error.  Also better --debug output.
When you merge &state and NULL, then most of the time you do want the
resulting state to be &undefined.  If you merge two known states then
you want the result to be &merged.

I also changed the --debug output a bit.  It now prints the ->possible
states after a merge.  I moved the debug output from merge_states() to the
callers.

Signed-off-by: Dan Carpenter <error27@gmail.com>