Handle declarations inside switch statement blocks.
This is a bit hacky.  I don't want that area to be unreachable so I removed
the nullify path.  But unfortunately states set there still aren't stored
past the first break statement.

Also you can declare variables in any compound statement not just inside
switch statements.  This is still not handled:

return;
{ int a;
foo:

According to smatch the line declaring "a" is unreachable.

Signed-off-by: Dan Carpenter <error27@gmail.com>