New function harmonize_states()
harmonize_states() makes sure that each implied true state matches an
implied false state and that they are different.

Setting the implied states currently erases the ->possible and ->pools
lists so it's a bad idea to do it for no reason.  Plus doing it for
no reason makes the --debug output longer.

If the there is only a true state and not a false state then when we or
the true and false states it gets set to &undefined.
&state + NULL => &undefined.

Signed-off-by: Dan Carpenter <error27@gmail.com>