Fix handling mutually exclusive states.
Mutually exclusive states can be valid.  For example:
if (!a && a = returns_nonnull())
The correct thing here is to say a is nonnull in the true path.

The old code said a was both isnull and nonnull which is wrong.  Also
the implementation was buggy so it just said that a was &merged without
saying which states were merged together to become &merged.

Signed-off-by: Dan Carpenter <error27@gmail.com>