Save all the pools a state goes through.
We need to have ->my_pools which is just a list of pools where the
state came from and also ->all_pools which is a list of all the pools
that a state has been through.

So to figure out the implied states for (a != NULL) we look at all the
->my_pools where a is nonnull.  All the states from those apply if they
have have the pool in their ->all-pools list.

Signed-off-by: Dan Carpenter <error27@gmail.com>