Free all the sm_states at the end of each function.
Most people probably think that it would be better to pay careful
attention to memory allocations and free them as you go along in
an organized way.  It turns out that method is hard to do.

I had to fix how the ->possible states were stored to avoid some
double frees and use after frees.

There are still lots of leaks of state_lists.

Also there is still a bug that triggers the OOM killer...

Signed-off-by: Dan Carpenter <error27@gmail.com>