Fix string memory leaks.
Internally smatch now makes a copy of the string so that the check script
can now free their own memory and internally smatch can free it's own
memory.

Also the function to free all the ->possibles and ->pools wasn't working.
That is now fixed.

There are still some memory leaks in some of the check scripts that
allocate their own states.

Signed-off-by: Dan Carpenter <error27@gmail.com>