implied pools:  cleanup how implied pools are created.
Implied pools help us know where a state comes from.  When two branches
are merged we save the pool.  sm_states have an ->all_pools which records
every pool that a state is a part of.  The first time an sm_state goes
through a pool that is recorded in the ->my_pools.  When we merge
sm_states both the ->my_pools and ->all_pools from both sides are copied
to the new state.

This patch cleans up how that is implemented.

Signed-off-by: Dan Carpenter <error27@gmail.com>