Remove unneeded code.
With the way that we currently set up the ->my_slists if we have something
like:
a = returns_nonnull();
if (x) {
...
}
At the end of the if statement we add the 'a' to different implied_pools
so is_really_same() returns false.  It needs to be rethought.

Anyway, the point is that this code makes no difference.

is_really_same() is used in one other place and for that place it often
returns true so it maybe is worthwhile.

Signed-off-by: Dan Carpenter <error27@gmail.com>