Fix __label__ handling
a) __label__ in gcc is not a type, it's a statement.  Accepted in the beginning
of compound-statement, has form __label__ ident-list;
b) instead of crapping into NS_SYMBOL namespace (and consequent shadowing
issues), reassign the namespace to NS_LABEL after we'd bound it.  We'll get
block scope and label namespace, i.e. what we get in gcc.
c) MOD_LABEL can be dropped now.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Christopher Li <sparse@chrisli.org>