implied:  handle &undefined better
foo = NULL;
x = 0;
if (bar) {
    foo = frob();
    x = 1;
}
Setting foo = frob() makes it &undefined but we still know that when foo is
non zero x == 1.

Also set things in smatch_extra to be &undefined earlier.  This is not a
complete solution because it doesn't deal with globals or members of
structures passed as arguments.  The way to deal with that is to set those
paths retroactively to &undefined in merge_slist().

Signed-off-by: Dan Carpenter <error27@gmail.com>