handle known conditions better
Conditions are known maybe because it is an if (1) for certain configs, or
maybe certain paths are not used on the first iteration through a loop, or
maybe it's just sloppiness.  Anyway, it leads to false positives if you
don't handle it.

This patch makes if statements not branch if the condition is known.  Also
it only checks the appropriate code path ->if_true or ->if_false.  I've
also made the known_condition_true/false() functions a bit cleverer.

Signed-off-by: Dan Carpenter <error27@gmail.com>