core:  add_unmatched_state_hook()
When we merge_slist() it used to be that a state would exist in on one
side but not on the other.  The merge hook could figure out what the
resulting state was.  The new way is for the unmatched state hook to say
what the state is on the side without a state, we set it there and then
merge the states.

The advantage to this way is that it sets up the ->my_list more accurately
and improves the implications.

Signed-off-by: Dan Carpenter <error27@gmail.com>