core:  improve anding condition handling.
The handling of merged states was obviously wrong before and so this is
an improvement.

It's a bit tricky because sometimes mutually exclusive states are
legit.
if (!foo && !(foo = kmalloc())) {

Sometimes when it complains about mutually exclusive states that
really does mean a condition is always false.  We set a lot of states
all the time because of implications.  I may remove the
"mutually exclusive" print statement again.

Signed-off-by: Dan Carpenter <error27@gmail.com>