check_memory: read freeing functions from file.
The check_frees_argument can be used to generate a list of functions
which free their argument.  Use this list instead of simply kfree by
itself.

Also it now prints an error if freed memory gets dereferenced.  The
downside is that adding this makes the check slower than I'd like.

Signed-off-by: Dan Carpenter <error27@gmail.com>