new (disabled) script:  check_leaks.c
check_leaks() looks for local variables that aren't freed on the error
path.  It turns out not to be as useful as desired and I think that is
because so many other static analysis tools have found this type of bug.
Probably if we used  a different allocation function besides kmalloc()
it would find more useful bugs.

Signed-off-by: Dan Carpenter <error27@gmail.com>