Update the gen_frees_list.sh and the generated data file.
The new data file has entries removed because it is using a newer version
of the check_frees_argument.c file.  Also it's updated to 2.6.29.  The
kernel.frees_argument.remove list is for functions that cause false
positives.  There were 2 functions called free_cell and one frees arguments
and the other one doesn't.

Signed-off-by: Dan Carpenter <error27@gmail.com>