Add check_type.c (disabled)
You are supposed to use kfree_skb() to free sk_buffs not kfree().  There
was a bug to do with this in a 2.6.29-rc kernel.  Check type.c checks for
these kinds of bugs.

The script finds the bug which was fixed in the final kernel but it doesn't
find any other bugs besides this.  It's disabled by default.

Signed-off-by: Dan Carpenter <error27@gmail.com>