Major memory saving.
These days clone_slist() can be just a pointer to sm_states instead of a
complete copy of all the sm_states.  Before it wasn't possible to do that
because we used to change the ->state member but now we just allocate a
new state.

Not making complete copies of everything all the time is a big win.

Signed-off-by: Dan Carpenter <error27@gmail.com>