Don't the implications for the state determining the implications.
When we set a state that implies that the current state is the implied
state for now.  So when we come to a branch all the check scripts
get to set their states.  Then anything that doesn't get set maybe has
implications that can be set.

The problem is for smatch_extra we can't set the state first or we
erase any implications.  But we also don't want to set the implications,
we want to set it ourselves.

So we smatch_implied calls the smatch_extra condition hook after it sets
the implications.  Also it doesn't set the states in smatch_implied.

Signed-off-by: Dan Carpenter <error27@gmail.com>