Revert "Don't the implications for the state determining the implications."
This reverts commit 4469500cba9be84b44cf1dc332b120ad133c5d2b.

The original patch made it so that if you had:
if (x == 1) {
The x == 1 would get set by smatch extra instead of smatch implied.  I
do not recall why that behavior was prefered.  I guess it fixed a bug.
I would think it would be better to set in smatch implied and preserve
the ->my_slists.

Anyway, I'm currently working on switch statements and I need smatch
implied to set it for that.