potential bug fix for gotos with implications
I can't think how to write a test case for this, but I think it is a bug
fix.

With gotos and switch statements we have to call clone_slist_and_states().
But say we clone a merged state, then decide that the unmerged state is
implied.  It's the same as if we didn't clone the states.  It probably
causes the roughly the same bugs.

Anyway, I can't see a downside to cloning the state here except that it
uses more memory.

we already do this for implications inside switch statements so that
doesn't  need to be modified.

Signed-off-by: Dan Carpenter <error27@gmail.com>