Rework implications to use ->pre_left and ->pre_right.
What I want eventually is for every sm_state to only have one pool.

We have this tree with merged sm_states as branches and unmerged sm_states
as leaves.  Then when we have implications we sort of create phantom
branches.  With this change we create one phantom branch for each real
branch.

Later on, hopefully when we merge two trees together we will be able to
replace the phantom branches with real branches sometimes.  This will
save memory.

Signed-off-by: Dan Carpenter <error27@gmail.com>